+++ xbmc/addons/kodi-dev-kit/include/kodi/addon-instance/Screensaver.h
@@ -9,68 +9,10 @@
 #pragma once
 
 #include "../AddonBase.h"
+#include "../c-api/addon-instance/screensaver.h"
 #include "../gui/renderHelper.h"
 
-extern "C"
-{
-
-struct AddonInstance_Screensaver;
-
-/*!
- * @brief Screensaver properties
- *
- * Not to be used outside this header.
- */
-typedef struct AddonProps_Screensaver
-{
-  void *device;
-  int x;
-  int y;
-  int width;
-  int height;
-  float pixelRatio;
-  const char *name;
-  const char *presets;
-  const char *profile;
-} AddonProps_Screensaver;
-
-/*!
- * @brief Screensaver callbacks
- *
- * Not to be used outside this header.
- */
-typedef struct AddonToKodiFuncTable_Screensaver
-{
-  KODI_HANDLE kodiInstance;
-} AddonToKodiFuncTable_Screensaver;
-
-/*!
- * @brief Screensaver function hooks
- *
- * Not to be used outside this header.
- */
-typedef struct KodiToAddonFuncTable_Screensaver
-{
-  KODI_HANDLE addonInstance;
-  bool (__cdecl* Start) (AddonInstance_Screensaver* instance);
-  void (__cdecl* Stop) (AddonInstance_Screensaver* instance);
-  void (__cdecl* Render) (AddonInstance_Screensaver* instance);
-} KodiToAddonFuncTable_Screensaver;
-
-/*!
- * @brief Screensaver instance
- *
- * Not to be used outside this header.
- */
-typedef struct AddonInstance_Screensaver
-{
-  AddonProps_Screensaver* props;
-  AddonToKodiFuncTable_Screensaver* toKodi;
-  KodiToAddonFuncTable_Screensaver* toAddon;
-} AddonInstance_Screensaver;
-
-} /* extern "C" */
-
+#ifdef __cplusplus
 namespace kodi
 {
 namespace addon
@@ -463,3 +405,4 @@ namespace addon
 
 } /* namespace addon */
 } /* namespace kodi */
+#endif /* __cplusplus */
